Replaced xsd with the Score Batcher version of it.Regenerated the code from the xsd.Replaced ETS TRT files with the simulated version of the TDS TRT files.Removed JSON/CSV support.
