+++ model/src/main/java/rdw/model/ObjectFactory.java
@@ -1,19 +1,22 @@
+
 package rdw.model;
 
 import javax.xml.bind.annotation.XmlRegistry;
 
 
 /**
- * This object contains factory methods for each Java content interface and Java element interface
- * generated in the rdw.model package.
- * <p>An ObjectFactory allows you to programatically construct new instances of the Java representation
- * for XML content. The Java representation of XML content can consist of schema derived interfaces
- * and classes representing the binding of schema type definitions, element declarations and model
- * groups.  Factory methods for each of these are provided in this class.
- *
- * This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802
- * See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a>
- * Generated on: 2016.12.14 at 10:13:14 AM PST
+ * This object contains factory methods for each 
+ * Java content interface and Java element interface 
+ * generated in the rdw.model package. 
+ * <p>An ObjectFactory allows you to programatically 
+ * construct new instances of the Java representation 
+ * for XML content. The Java representation of XML 
+ * content can consist of schema derived interfaces 
+ * and classes representing the binding of schema 
+ * type definitions, element declarations and model 
+ * groups.  Factory methods for each of these are 
+ * provided in this class.
+ * 
  */
 @XmlRegistry
 public class ObjectFactory {
@@ -43,51 +46,59 @@ public class ObjectFactory {
     }
 
     /**
-     * Create an instance of {@link ToolUsage }
+     * Create an instance of {@link TDSReport.ToolUsage }
      * 
      */
-    public ToolUsage createTDSReportToolUsage() {
-        return new ToolUsage();
+    public TDSReport.ToolUsage createTDSReportToolUsage() {
+        return new TDSReport.ToolUsage();
     }
 
     /**
-     * Create an instance of {@link Opportunity }
+     * Create an instance of {@link TDSReport.Opportunity }
      * 
      */
-    public Opportunity createTDSReportOpportunity() {
-        return new Opportunity();
+    public TDSReport.Opportunity createTDSReportOpportunity() {
+        return new TDSReport.Opportunity();
     }
 
     /**
-     * Create an instance of {@link Opportunity.Item }
+     * Create an instance of {@link TDSReport.Opportunity.Item }
      * 
      */
-    public Opportunity.Item createTDSReportOpportunityItem() {
-        return new Opportunity.Item();
+    public TDSReport.Opportunity.Item createTDSReportOpportunityItem() {
+        return new TDSReport.Opportunity.Item();
     }
 
     /**
-     * Create an instance of {@link Examinee }
+     * Create an instance of {@link TDSReport.Examinee }
      * 
      */
-    public Examinee createTDSReportExaminee() {
-        return new Examinee();
+    public TDSReport.Examinee createTDSReportExaminee() {
+        return new TDSReport.Examinee();
     }
 
     /**
-     * Create an instance of {@link Test }
+     * Create an instance of {@link TDSReport.Test }
      * 
      */
-    public Test createTDSReportTest() {
-        return new Test();
+    public TDSReport.Test createTDSReportTest() {
+        return new TDSReport.Test();
     }
 
     /**
-     * Create an instance of {@link Comment }
+     * Create an instance of {@link TDSReport.Comment }
      * 
      */
-    public Comment createTDSReportComment() {
-        return new Comment();
+    public TDSReport.Comment createTDSReportComment() {
+        return new TDSReport.Comment();
+    }
+
+    /**
+     * Create an instance of {@link ScoreInfoType.ScoreRationale }
+     * 
+     */
+    public ScoreInfoType.ScoreRationale createScoreInfoTypeScoreRationale() {
+        return new ScoreInfoType.ScoreRationale();
     }
 
     /**
@@ -99,67 +110,67 @@ public class ObjectFactory {
     }
 
     /**
-     * Create an instance of {@link ToolUsage.ToolPage }
+     * Create an instance of {@link TDSReport.ToolUsage.ToolPage }
      * 
      */
-    public ToolUsage.ToolPage createTDSReportToolUsageToolPage() {
-        return new ToolUsage.ToolPage();
+    public TDSReport.ToolUsage.ToolPage createTDSReportToolUsageToolPage() {
+        return new TDSReport.ToolUsage.ToolPage();
     }
 
     /**
-     * Create an instance of {@link Opportunity.Segment }
+     * Create an instance of {@link TDSReport.Opportunity.Segment }
      * 
      */
-    public Opportunity.Segment createTDSReportOpportunitySegment() {
-        return new Opportunity.Segment();
+    public TDSReport.Opportunity.Segment createTDSReportOpportunitySegment() {
+        return new TDSReport.Opportunity.Segment();
     }
 
     /**
-     * Create an instance of {@link Opportunity.Accommodation }
+     * Create an instance of {@link TDSReport.Opportunity.Accommodation }
      * 
      */
-    public Opportunity.Accommodation createTDSReportOpportunityAccommodation() {
-        return new Opportunity.Accommodation();
+    public TDSReport.Opportunity.Accommodation createTDSReportOpportunityAccommodation() {
+        return new TDSReport.Opportunity.Accommodation();
     }
 
     /**
-     * Create an instance of {@link Opportunity.Score }
+     * Create an instance of {@link TDSReport.Opportunity.Score }
      * 
      */
-    public Opportunity.Score createTDSReportOpportunityScore() {
-        return new Opportunity.Score();
+    public TDSReport.Opportunity.Score createTDSReportOpportunityScore() {
+        return new TDSReport.Opportunity.Score();
     }
 
     /**
-     * Create an instance of {@link Opportunity.GenericVariable }
+     * Create an instance of {@link TDSReport.Opportunity.GenericVariable }
      * 
      */
-    public Opportunity.GenericVariable createTDSReportOpportunityGenericVariable() {
-        return new Opportunity.GenericVariable();
+    public TDSReport.Opportunity.GenericVariable createTDSReportOpportunityGenericVariable() {
+        return new TDSReport.Opportunity.GenericVariable();
     }
 
     /**
-     * Create an instance of {@link Opportunity.Item.Response }
+     * Create an instance of {@link TDSReport.Opportunity.Item.Response }
      * 
      */
-    public Opportunity.Item.Response createTDSReportOpportunityItemResponse() {
-        return new Opportunity.Item.Response();
+    public TDSReport.Opportunity.Item.Response createTDSReportOpportunityItemResponse() {
+        return new TDSReport.Opportunity.Item.Response();
     }
 
     /**
-     * Create an instance of {@link Examinee.ExamineeAttribute }
+     * Create an instance of {@link TDSReport.Examinee.ExamineeAttribute }
      * 
      */
-    public Examinee.ExamineeAttribute createTDSReportExamineeExamineeAttribute() {
-        return new Examinee.ExamineeAttribute();
+    public TDSReport.Examinee.ExamineeAttribute createTDSReportExamineeExamineeAttribute() {
+        return new TDSReport.Examinee.ExamineeAttribute();
     }
 
     /**
-     * Create an instance of {@link Examinee.ExamineeRelationship }
+     * Create an instance of {@link TDSReport.Examinee.ExamineeRelationship }
      * 
      */
-    public Examinee.ExamineeRelationship createTDSReportExamineeExamineeRelationship() {
-        return new Examinee.ExamineeRelationship();
+    public TDSReport.Examinee.ExamineeRelationship createTDSReportExamineeExamineeRelationship() {
+        return new TDSReport.Examinee.ExamineeRelationship();
     }
 
 }
