+++ model/src/main/java/rdw/model/ScoreInfoType.java
@@ -1,3 +1,4 @@
+
 package rdw.model;
 
 import java.util.ArrayList;
@@ -22,7 +23,17 @@ import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="ScoreRationale" type="{http://www.w3.org/2001/XMLSchema}anyType" minOccurs="0"/>
+ *         &lt;element name="ScoreRationale" minOccurs="0">
+ *           &lt;complexType>
+ *             &lt;complexContent>
+ *               &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+ *                 &lt;sequence>
+ *                   &lt;element name="Message" type="{http://www.w3.org/2001/XMLSchema}anyType" minOccurs="0"/>
+ *                 &lt;/sequence>
+ *               &lt;/restriction>
+ *             &lt;/complexContent>
+ *           &lt;/complexType>
+ *         &lt;/element>
  *         &lt;element name="SubScoreList" minOccurs="0">
  *           &lt;complexType>
  *             &lt;complexContent>
@@ -36,6 +47,7 @@ import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
  *         &lt;/element>
  *       &lt;/sequence>
  *       &lt;attribute name="scorePoint" type="{}UFloatAllowNegativeOne" />
+ *       &lt;attribute name="maxScore" type="{}UFloatAllowNegativeOne" />
  *       &lt;attribute name="scoreDimension" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
  *       &lt;attribute name="scoreStatus">
  *         &lt;simpleType>
@@ -47,14 +59,7 @@ import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
  *           &lt;/restriction>
  *         &lt;/simpleType>
  *       &lt;/attribute>
- *       &lt;attribute name="conditionCode">
- *         &lt;simpleType>
- *           &lt;restriction base="{http://www.w3.org/2001/XMLSchema}string">
- *             &lt;pattern value="(|[A-Z])"/>
- *             &lt;maxLength value="1"/>
- *           &lt;/restriction>
- *         &lt;/simpleType>
- *       &lt;/attribute>
+ *       &lt;attribute name="confLevel" type="{http://www.w3.org/2001/XMLSchema}anySimpleType" />
  *     &lt;/restriction>
  *   &lt;/complexContent>
  * &lt;/complexType>
@@ -70,29 +75,32 @@ import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 public class ScoreInfoType {
 
     @XmlElement(name = "ScoreRationale")
-    protected Object scoreRationale;
+    protected ScoreInfoType.ScoreRationale scoreRationale;
     @XmlElement(name = "SubScoreList")
     protected ScoreInfoType.SubScoreList subScoreList;
     @XmlAttribute(name = "scorePoint")
     protected String scorePoint;
+    @XmlAttribute(name = "maxScore")
+    protected String maxScore;
     @XmlAttribute(name = "scoreDimension")
     @XmlSchemaType(name = "anySimpleType")
     protected String scoreDimension;
     @XmlAttribute(name = "scoreStatus")
     @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
     protected String scoreStatus;
-    @XmlAttribute(name = "conditionCode")
-    protected String conditionCode;
+    @XmlAttribute(name = "confLevel")
+    @XmlSchemaType(name = "anySimpleType")
+    protected String confLevel;
 
     /**
      * Gets the value of the scoreRationale property.
      * 
      * @return
      *     possible object is
-     *     {@link Object }
+     *     {@link ScoreInfoType.ScoreRationale }
      *     
      */
-    public Object getScoreRationale() {
+    public ScoreInfoType.ScoreRationale getScoreRationale() {
         return scoreRationale;
     }
 
@@ -101,10 +109,10 @@ public class ScoreInfoType {
      * 
      * @param value
      *     allowed object is
-     *     {@link Object }
+     *     {@link ScoreInfoType.ScoreRationale }
      *     
      */
-    public void setScoreRationale(Object value) {
+    public void setScoreRationale(ScoreInfoType.ScoreRationale value) {
         this.scoreRationale = value;
     }
 
@@ -157,6 +165,30 @@ public class ScoreInfoType {
     }
 
     /**
+     * Gets the value of the maxScore property.
+     * 
+     * @return
+     *     possible object is
+     *     {@link String }
+     *     
+     */
+    public String getMaxScore() {
+        return maxScore;
+    }
+
+    /**
+     * Sets the value of the maxScore property.
+     * 
+     * @param value
+     *     allowed object is
+     *     {@link String }
+     *     
+     */
+    public void setMaxScore(String value) {
+        this.maxScore = value;
+    }
+
+    /**
      * Gets the value of the scoreDimension property.
      * 
      * @return
@@ -205,27 +237,82 @@ public class ScoreInfoType {
     }
 
     /**
-     * Gets the value of the conditionCode property.
+     * Gets the value of the confLevel property.
      * 
      * @return
      *     possible object is
      *     {@link String }
      *     
      */
-    public String getConditionCode() {
-        return conditionCode;
+    public String getConfLevel() {
+        return confLevel;
     }
 
     /**
-     * Sets the value of the conditionCode property.
+     * Sets the value of the confLevel property.
      * 
      * @param value
      *     allowed object is
      *     {@link String }
      *     
      */
-    public void setConditionCode(String value) {
-        this.conditionCode = value;
+    public void setConfLevel(String value) {
+        this.confLevel = value;
+    }
+
+
+    /**
+     * <p>Java class for anonymous complex type.
+     * 
+     * <p>The following schema fragment specifies the expected content contained within this class.
+     * 
+     * <pre>
+     * &lt;complexType>
+     *   &lt;complexContent>
+     *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
+     *       &lt;sequence>
+     *         &lt;element name="Message" type="{http://www.w3.org/2001/XMLSchema}anyType" minOccurs="0"/>
+     *       &lt;/sequence>
+     *     &lt;/restriction>
+     *   &lt;/complexContent>
+     * &lt;/complexType>
+     * </pre>
+     * 
+     * 
+     */
+    @XmlAccessorType(XmlAccessType.FIELD)
+    @XmlType(name = "", propOrder = {
+        "message"
+    })
+    public static class ScoreRationale {
+
+        @XmlElement(name = "Message")
+        protected Object message;
+
+        /**
+         * Gets the value of the message property.
+         * 
+         * @return
+         *     possible object is
+         *     {@link Object }
+         *     
+         */
+        public Object getMessage() {
+            return message;
+        }
+
+        /**
+         * Sets the value of the message property.
+         * 
+         * @param value
+         *     allowed object is
+         *     {@link Object }
+         *     
+         */
+        public void setMessage(Object value) {
+            this.message = value;
+        }
+
     }
