--- model/src/test/resources/asmt1.json
@@ -1,1 +1,0 @@
-{"Overall": {"MaxScore": 0, "MinScore": 0}, "content": "assessment", "PerformanceLevels": {"Level1": {"Name": "", "CutPoint": 0}, "Level2": {"Name": "", "CutPoint": ""}, "Level3": {"Name": "", "CutPoint": ""}, "Level4": {"Name": "", "CutPoint": ""}, "Level5": {"Name": "", "CutPoint": ""}}, "Identification": {"Type": "INTERIM ASSESSMENT BLOCKS", "Subject": "ELA", "Guid": "0012378f0ab34e22a07b495edafeef", "EffectiveDate": "20170101", "Period": "2017-01-01", "Year": 2017, "Version": "V1"}, "Claims": {"Claim4": {"MaxScore": "", "Name": "", "MinScore": ""}, "Claim2": {"MaxScore": 0, "Name": "Grade 5", "MinScore": 0}, "Claim3": {"MaxScore": 0, "Name": "", "MinScore": 0}, "Claim1": {"MaxScore": 2400, "Name": "Opinion Performance Task", "MinScore": 1200}}, "ClaimsPerformanceLevel": {"Level1": {"Name": "Below Standard"}, "Level2": {"Name": "At/Near Standard"}, "Level3": {"Name": "Above Standard"}}}
