--- model/src/test/resources/sr1.json
@@ -1,1 +1,0 @@
-{"Source": {"TestRegCallbackURL": "SateTestReg.gov/StuReg/CallBack", "TestRegSysID": "c458fbd4-07d9-4d8d-927f-ee2fb11b6963"}, "Identification": {"AcademicYear": 2015, "Guid": "9940d0ed-6ed0-4be5-b6cb-d1278acecfda", "ExtractDate": "2014-02-27"}, "Content": "StudentRegistration"}
